go tree impl of balanced tree
go tree impl of balanced tree
